xrandr-scaling: Trigger a monitors rebuild only if scale changed
When in logical global UI scale, and we change the scale from an integer value
to another integer value, we have de-facto to only change the UI scale, while
no crtc configuration has to be done.

So, avoid applying the crtc assignments in such case (as they wouldn't lead to
any xserver reconfiguration and RRScreenChangeNotify event anyway), but update
the logical configuration that will eventually lead to an UI scale factor
update and related signaling.

Fixes LP: #1823485