debian/patches: Sync master clock timelines to hardware vsync
Improves the animations smoothness as it uses the hardware vsync to trigger
them instead of timing values

The MR was OK'ed to be merged by upstream, but the reason was not merged is
that there was some ongoing refactor that was supposed to land for 3.36.
Given the refactor didn't land yet, is just better to have this on board.

https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/724