xrandr-scaling: Support shell direct-rendering inhibitions
When the shell inhibits the direct rendering we were ignoring the requests
if using a scaled monitor, however this won't permit to access the shell
widgets (i.e. overview, alt/tab) while we're using direct rendering.

Fixes this by disabling the monitor scaling only if it was previously set
and by disabling it only if there's no any possible direct-rendering pick.

LP: #1862081