xrandr-scaling: Take care of global UI scale when restoring from config
When we restore the configuration from a local file we need to properly
derive the logical monitor sizes taking care of the global UI scale value
we're going to apply, otherwise we won't be able to restore the previous
configuration as per non-adjacent monitors.

To do this, multiply all the monitors sizes per the global UI scale value
and downscale it again when verifying if the logical mode can be applied
with any known mode.

As per this, also properly verify configurations that are provided using the
dbus API.

LP: #1825593