xrandr-scaling: Use pre-computed crtc layout size for computing screen size
Don't manually compute the unscaled display size taking care of its rotation
as this is now already happening at construction time.
And take care of the configured scaling to compute its scaling value.

This was causing a regressions, making vertical monitors not to be properly
configured (under nvidia) and fractional setups not to be properly defined
as they were using an invalid screen size.

Also update the CRTCs layout configuration when using UI-down scaling as in
such case we need to scale the saved values.
And ensure that we multiply the computed CRTC scale by the UI size if the
resource scale atom is not set.

LP: #1874567
LP: #1873403