xrandr-scaling: Apply proper layout monitor config when using fractional scaling
Fractional scaling layout is ignored when using configurations with a GPU
that has not hotplug mode update.

In such case we need to adapt the configuration to the layout we want to use
otherwise we won't apply the proper configuration.

We should also try to adapt applied configuration to requested layout, when
picking a saved one.

Even previous commit was for the same bugggg!

LP: #1884754