debian/control: Do not break on old gnome-shell versions
This was included as part of commit 5be177fb (and wrongly updated as
part of d68efdd) to ensure that a libgnome-desktop compatible shell was
used, so we should not break again on previous version as we don't now.

However it's not needed to keep the break at all at this point, given
that the only mutter-8 compatible gnome-shell will be one that will only
depend on the same libgnome-desktop and so there's no risk that
mismatching versions gets ever pulled by the shell and mutter.