debian: Enable sysprof profiler on linux
mutter will link statically with libsysprof-capture now so no new
shared dependencies will be added, thus this won't bring in the sysprof
dependencies (that include UI libraries that are unneeded by mutter).

Even though upstream sources only require sysprof 3.37.2, we depend on
sysprof 3.40.1 as this is the first version that won't cause its private
symbols to be exported as part of mutter too.