Stop patching in Xorg synaptics driver support
Upstream rejected this patch back in 2018, but we continued to apply it
as a transitional step for Ubuntu 18.04. Since then we've had stable/LTS
releases of Debian 10, Ubuntu 18.04, Debian 11 and Ubuntu 20.04,
libinput has improved, and both Debian and Ubuntu have switched to
running GNOME in Wayland mode by default; let's have another try at
following upstream on this.

Closes: #993171