d/rules: Use upstream driver selection, except on 32-bit ARM
Specifying -Ddefault_driver=gl on x86 (and other architectures that are
not 32-bit ARM) is no longer the same as the upstream default: the
upstream default is now gl3, meaning modern OpenGL, whereas the gl driver
is legacy (pre-3.2) OpenGL. Assume upstream are doing this for a reason.

For now we continue to specify -Ddefault_driver=gles2 on 32-bit ARM, and
apply the patch that adds the default_driver option, because 32-bit ARM
has historically been an exception to the rules.