debian/patches: Allow any drag timestamp as drag start serial
In some scenarios under wayland DnD may not work because the start-drag
serial is not matching the one we expect for the pointer that we set
when handling the button event.

This is because since commit 26676a82 we're repeatedly pinging the
windows at each event to ensure that they're still alive, but we're
using the event timestamp for that, and this makes wayland to use such
value as the start-drag serial.

To avoid this, let's keep track of all the grab timestamps we got while
dragging and accept a serial to match any of these.

LP: #1964541