Prefer GPUs with built-in panels connected when selecting a primary GPU
[ upstream commit ecbe4a5dada65ca8c753ddb1e9ed43257a89a2e1 ]

Some laptops don't have the integrated GPU come earlier in the PCI topology
which means that when mutter enumerates the GPUs it may select the dGPU
as the primary.

In a laptop context, this is unlikely to work because the eDP panel can
only be actively connected to a single GPU at a time.

Closes: https://gitlab.gnome.org/GNOME/mutter/-/issues/3746
Suggested-by: Jonas Ådahl
Signed-off-by: Mario Limonciello <mario.limonciello@amd.com>
Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4114>
Signed-off-by: Tim Crawford <tcrawford@system76.com>